+++ magda-web-client/src/Components/SearchFacets/Region.js
@@ -2,7 +2,7 @@ import { addRegion, resetRegion } from "../../actions/datasetSearchActions";
 import { connect } from "react-redux";
 import { fetchRegionSearchResults } from "../../actions/facetRegionSearchActions";
 import defined from "../../helpers/defined";
-import FacetRegion from "./FacetRegion";
+import RegionWrapper from "./RegionWrapper";
 import React, { Component } from "react";
 
 class Region extends Component {
@@ -24,13 +24,13 @@ class Region extends Component {
         this.props.closeFacet();
     }
 
-    onResetRegionFacet() {
+    async onResetRegionFacet() {
         this.props.updateQuery({
             regionId: undefined,
             regionType: undefined,
             page: undefined
         });
-        this.props.dispatch(resetRegion());
+        await this.props.dispatch(resetRegion());
         this.props.closeFacet();
     }
 
@@ -38,11 +38,16 @@ class Region extends Component {
         this.props.dispatch(fetchRegionSearchResults(facetKeyword));
     }
 
+    myMeth = () => {
+        this.props.dispatch(resetRegion());
+    };
+
     render() {
         return (
-            <FacetRegion
+            <RegionWrapper
                 title="location"
                 id="region"
+                myMeth={this.myMeth}
                 hasQuery={
                     defined(this.props.activeRegion.regionType) &&
                     defined(this.props.activeRegion.regionId)